pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel", "setuptools_scm[toml]>=6.2"]
build-backend = "setuptools.build_meta"

[project]
name = "streamlitforge"
version = "0.1.0"
description = "AI-Powered Streamlit Application Builder CLI"
readme = "README.md"
authors = [
    {name = "StreamlitForge Team", email = "team@streamlitforge.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["streamlit", "cli", "ai", "llm", "scaffolding"]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "jinja2>=3.1.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "numpy>=1.21.0",
    "streamlit>=1.28.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=22.0.0",
    "flake8>=5.0.0",
    "mypy>=0.990",
]

[project.urls]
Homepage = "https://github.com/yourusername/streamlitforge"
Documentation = "https://streamlitforge.readthedocs.io"
Repository = "https://github.com/yourusername/streamlitforge"
"Bug Tracker" = "https://github.com/yourusername/streamlitforge/issues"

[project.scripts]
streamlitforge = "streamlitforge.cli:main"

[tool.setuptools]
packages = ["streamlitforge"]

[tool.setuptools.package-data]
streamlitforge = ["templates/**/*", "patterns/**/*"]

[tool.pytest.ini_options]
testpaths = ["streamlitforge/tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311']
include = '\.pyi?$'

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
